KNL-1587 - Moving Exception class
(cherry picked from commit 3535c9f6bfacc858748044ffb72c3bec375c29e2)
