+++ kernel/kernel-impl/src/main/java/org/sakaiproject/authz/impl/SakaiSecurity.java
@@ -34,6 +34,7 @@ import org.sakaiproject.entity.api.Reference;
 import org.sakaiproject.event.api.Event;
 import org.sakaiproject.event.api.EventTrackingService;
 import org.sakaiproject.exception.IdUnusedException;
+import org.sakaiproject.exception.IllegalSecurityAdvisorException;
 import org.sakaiproject.memory.api.Cache;
 import org.sakaiproject.memory.api.MemoryService;
 import org.sakaiproject.site.api.Site;
